Don't call `mkdirs()` unnecessarily
